feat: thêm chức năng quản lý người dùng và vai trò
- Tạo mới các controller cho user và role, bao gồm UserController và RoleController.
- Thêm các phương thức để lấy danh sách người dùng, tạo, cập nhật và xóa người dùng.
- Cập nhật AdminInitializer để thiết lập vai trò và người dùng mặc định trong cơ sở dữ liệu.
- Cải thiện LevelService với các phương thức mới để lấy level mặc định cho người dùng và lấy level theo loại.
- Thêm các kiểu dữ liệu mới cho request và response user.
- Cải thiện giao diện người dùng với các modal cho việc thêm và xóa người dùng.
- Cập nhật các thành phần giao diện để hỗ trợ hiển thị thông tin người dùng và vai trò.

+++ backend-comic/src/main/java/com/thjvjpxx/backend_comic/model/Level.java
@@ -43,7 +43,7 @@ public class Level {
     String color;
 
     @Column(name = "exp_required")
-    double expRequired;
+    Integer expRequired;
 
     @Column(name = "url_gif")
     String urlGif;
